api/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.1
celery==5.4.0
redis==5.0.7
openai==1.43.0
supabase==2.6.0
python-dotenv==1.0.1
pydantic==2.9.2
